+++ R/RcppExports.R
@@ -173,3 +173,7 @@
     .Call(`_RcppCWB_cl_load_corpus`, corpus, registry_dir)
 }
 
+.cl_corpora <- function() {
+    .Call(`_RcppCWB_cl_corpora`)
+}
+